move more message wrapper stuff to message, merge CallMessage into Message this enables the context menu for callmessage
